Moved the alter table nodes into alter_table.go
