+++ sql/analyzer/rules.go
@@ -27,7 +27,6 @@ var OnceBeforeDefault = []Rule{
 	{"resolve_tables", resolveTables},
 	{"resolve_drop_constraint", resolveDropConstraint},
 	{"validate_drop_constraint", validateDropConstraint},
-	{"validate_drop_constraint", validateDropConstraint},
 	{"load_check_constraints", loadChecks},
 	{"resolve_set_variables", resolveSetVariables},
 	{"resolve_create_like", resolveCreateLike},
